+++ src/chat/frontend/chat-root/components/chat-modal/memory-modal/memory-modal.ts
@@ -3,7 +3,7 @@ import { customElement, property, query, state } from "lit/decorators.js";
 
 import type { MemoryScope, MemoryViewEntry } from "../../../../../shared";
 import { themeIcons } from "../../../../styles";
-import "../../../../template-components/button-box";
+import "../../../../template-components/button";
 import "../../../../template-components/button-row";
 import { BaseModal } from "../../../../template-components/modal/base-modal";
 import "../../../../template-components/text-box";