+++ src/chat/frontend/template-components/button/box-button/index.ts
@@ -1,0 +1,1 @@
+export { AcceptButton, AcceptAllButton, AddButton, ButtonBox, CancelButton, SendButton } from "./box-button";